code refactor, setup global context
