+++ src/App.js
@@ -1,106 +1,19 @@
-import React, { useState, useEffect } from "react";
-import { FaSearch } from "react-icons/fa";
-import Photo from "./Photo";
-import Loader from "react-loader-spinner";
-
-const clientID = `?client_id=${process.env.REACT_APP_ACCESS_KEY}`;
-const mainUrl = `https://api.unsplash.com/photos/`;
-const searchUrl = `https://api.unsplash.com/search/photos/`;
+import React from "react";
+import SearchForm from "./components/SearchForm";
+import Photos from "./components/Photos";
+import { useAppContext } from "./context/context";
 
 function App() {
-  const [loading, setLoading] = useState(false);
-  const [photos, setPhotos] = useState([]);
-  const [page, setPage] = useState(1);
-  const [query, setQuery] = useState("");
+  const { query, handleSubmit, setQuery, photos, loading } = useAppContext();
 
-  const fetchImages = async () => {
-    let url;
-    const urlPage = `&page=${page}`;
-    const urlQuery = `&query=${query}`;
-    if (query) {
-      url = `${searchUrl}${clientID}${urlPage}${urlQuery}`;
-    } else {
-      url = `${mainUrl}${clientID}${urlPage}`;
-    }
-    try {
-      setLoading(true);
-      const response = await fetch(url);
-      const data = await response.json();
-
-      setPhotos((oldPhotos) => {
-        debugger;
-        if (query && page === 1) {
-          return data.results;
-        } else if (query) {
-          return [...oldPhotos, ...data.results];
-        } else {
-          return [...oldPhotos, ...data];
-        }
-      });
-      setLoading(false);
-    } catch (err) {
-      setLoading(false);
-      console.log(err);
-    }
-  };
-
-  useEffect(() => {
-    fetchImages();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [page]);
-
-  useEffect(() => {
-    const event = window.addEventListener("scroll", () => {
-      if (
-        (!loading && window.innerHeight + window.scrollY) >=
-        document.body.scrollHeight - 2
-      ) {
-        setPage((oldPage) => {
-          return oldPage + 1;
-        });
-      }
-    });
-    return () => window.removeEventListener("scroll", event);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    setPage(() => 1);
-    fetchImages();
-  };
   return (
     <main>
-      <section className="search">
-        <form className="search-form" onSubmit={handleSubmit}>
-          <input
-            type="text"
-            placeholder="search"
-            className="form-input"
-            value={query}
-            onChange={(e) => setQuery(e.target.value)}
-          />
-          <button type="submit" className="submit-btn">
-            <FaSearch />
-          </button>
-        </form>
-      </section>
-      <section className="photos">
-        <div className="photos-center">
-          {photos.map((photo, index) => {
-            return <Photo key={index} {...photo} />;
-          })}
-          {loading && (
-            <Loader
-              type="Oval"
-              color="#617e98"
-              height={80}
-              width={80}
-              className="loading"
-            />
-          )}
-        </div>
-      </section>
+      <SearchForm
+        query={query}
+        handleSubmit={handleSubmit}
+        setQuery={setQuery}
+      />
+      <Photos photos={photos} loading={loading} />
     </main>
   );
 }
