+++ src/context/context.js
@@ -1,0 +1,91 @@
+import React, { useContext, useState, useEffect } from "react";
+
+const clientID = `?client_id=${process.env.REACT_APP_ACCESS_KEY}`;
+const mainUrl = `https://api.unsplash.com/photos/`;
+const searchUrl = `https://api.unsplash.com/search/photos/`;
+
+const AppContext = React.createContext();
+
+function AppProvider({ children }) {
+  const [loading, setLoading] = useState(false);
+  const [photos, setPhotos] = useState([]);
+  const [page, setPage] = useState(1);
+  const [query, setQuery] = useState("");
+
+  const fetchImages = async () => {
+    let url;
+    const urlPage = `&page=${page}`;
+    const urlQuery = `&query=${query}`;
+    if (query) {
+      url = `${searchUrl}${clientID}${urlPage}${urlQuery}`;
+    } else {
+      url = `${mainUrl}${clientID}${urlPage}`;
+    }
+    try {
+      setLoading(true);
+      const response = await fetch(url);
+      const data = await response.json();
+
+      setPhotos((oldPhotos) => {
+        debugger;
+        if (query && page === 1) {
+          return data.results;
+        } else if (query) {
+          return [...oldPhotos, ...data.results];
+        } else {
+          return [...oldPhotos, ...data];
+        }
+      });
+      setLoading(false);
+    } catch (err) {
+      setLoading(false);
+      console.log(err);
+    }
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    setPage(() => 1);
+    fetchImages();
+  };
+
+  useEffect(() => {
+    fetchImages();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [page]);
+
+  useEffect(() => {
+    const event = window.addEventListener("scroll", () => {
+      if (
+        (!loading && window.innerHeight + window.scrollY) >=
+        document.body.scrollHeight - 2
+      ) {
+        setPage((oldPage) => {
+          return oldPage + 1;
+        });
+      }
+    });
+    return () => window.removeEventListener("scroll", event);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  return (
+    <AppContext.Provider
+      value={{
+        query,
+        handleSubmit,
+        setQuery,
+        photos,
+        loading,
+      }}
+    >
+      {children}
+    </AppContext.Provider>
+  );
+}
+
+function useAppContext() {
+  return useContext(AppContext);
+}
+
+export { AppProvider, useAppContext };
